% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{paretoDens}
\alias{paretoDens}
\title{The density of the generalized pareto distribution}
\usage{
paretoDens(x, location, scale, shape, log = TRUE)
}
\description{
The density of the generalized pareto distribution
}
