% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{paretoGrad}
\alias{paretoGrad}
\title{Computes the gradient of the minus log-likelihood of the generalized pareto distribution}
\usage{
paretoGrad(params, x, location = NULL, weights = NULL, barrier = 0.005)
}
\description{
Computes the gradient of the minus log-likelihood of the generalized pareto distribution
}
