% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{paretoLogLik}
\alias{paretoLogLik}
\title{Computes the log-likelihood of the generalized pareto distribution}
\usage{
paretoLogLik(params, x, location = NULL, weights = NULL, barrier = 0)
}
\description{
Computes the log-likelihood of the generalized pareto distribution
}
