% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{paretoML}
\alias{paretoML}
\title{Computes the MLE of a generalized pareto model}
\usage{
paretoML(x, location = NULL, weights = NULL, barrier = 0.005)
}
\description{
Computes the MLE of a generalized pareto model
}
