% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTruncNormMix.R
\name{paretoTruncNormMix}
\alias{paretoTruncNormMix}
\title{A function for fitting a Genearlized Pareto/Truncated Normal mixture to a data}
\usage{
paretoTruncNormMix(x, threshold, normComps = 1, iterations = 100,
  paretoComp = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{the observed (truncated) z-scores}

\item{threshold}{the threshold used for screening, the selection rule is abs(x) > abs(threshold)}

\item{normComps}{number of truncated normal components to fit}

\item{iterations}{number of EM iterations}

\item{paretoComp}{whether to include a Generalzed Pareto mixture component}

\item{verbose}{whether to print a progress bar}
}
\description{
The aim of this function is to estimate the marginal distribution of a normal, convolved
with an unknown prior and then truncated. The model fit can be used to compute the
Empirical Bayes rule using Tweedy's Formula.
}
\seealso{
\code{\link{predict.ptnMix}}
}
