% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{paretoTweedy}
\alias{paretoTweedy}
\title{Computes the tweedy correction based on an estimated generalized pareto distribution}
\usage{
paretoTweedy(x, location, scale, shape)
}
\description{
Computes the tweedy correction based on an estimated generalized pareto distribution
}
