% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{paretoTweedyCorrection}
\alias{paretoTweedyCorrection}
\title{Tweedy correction for tails of distributions based on the generalized pareto distribution}
\usage{
paretoTweedyCorrection(x, threshold, weights = NULL, barrier = 0.005)
}
\description{
Tweedy correction for tails of distributions based on the generalized pareto distribution
}
