% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoNormMix.R
\name{predict.normParetoTweed}
\alias{predict.normParetoTweed}
\title{Computes the tweedy estimate for selected quantities based on a fitted normal-pareto mixture model}
\usage{
\method{predict}{normParetoTweed}(object, ...)
}
\description{
Computes the tweedy estimate for selected quantities based on a fitted normal-pareto mixture model
}
