% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{predict.paretoTweed}
\alias{predict.paretoTweed}
\title{Compute tweedy correction based on estimated Generalized Pareto Marginal Distribution}
\usage{
\method{predict}{paretoTweed}(object, newX = NULL, ...)
}
\description{
Compute tweedy correction based on estimated Generalized Pareto Marginal Distribution
}
