% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asafGmodeling.R
\name{predict.truncDeconv}
\alias{predict.truncDeconv}
\title{Computes the Bayes Rule based on a \code{truncDeconv} model fit}
\usage{
\method{predict}{truncDeconv}(object, binned = FALSE, newX = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{\link{truncDeconv}}}

\item{binned}{whether to return the binned Bayes rule, or the linear interpolation
for the data/newX}

\item{newX}{an optional new dataset to compute the Empirical Bayes Rule for}
}
\description{
Computes the Bayes rule for \code{\link{truncDeconv}} model fit.
}
