% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asafGmodeling.R
\name{truncDeconv}
\alias{truncDeconv}
\title{Computes Empirical Bayes Estimates via Deconvolution for Truncated Observations}
\usage{
truncDeconv(x, threshold, meanValues = NULL, twoSided = TRUE,
  splineDegree = 10, binWidth = 0.1, nBins = 100)
}
\arguments{
\item{x}{the observed (truncated) sample of z-scores}

\item{threshold}{the threshold used to screen the observations}

\item{meanValues}{a grid of discrete mean parameter values}

\item{twoSided}{whether the selection was one-sided or two-sided.
if TRUE, then a selection rule of abs(x) > abs(threshold) is assumed. Otherwise,
a selection rule of x > threshold is assumed.}

\item{splineDegree}{the number of degrees for the spline fit in the deconvolution}

\item{binWidth}{the function discretizes the sample into bins, this is the bin width.
Take precedence over nBins.}

\item{nBins}{number of bins to use in discretization}
}
\description{
This function has a similar functionality as \code{\link[deconvolveR]{deconv}}, excpet
that it is meant to handle the case in which only the selected are observed.
}
\seealso{
\code{\link{predict.truncDeconv}}
}
