% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{truncLogNormDens}
\alias{truncLogNormDens}
\title{Log-likelihood of the (two sided) truncated normal distribution}
\usage{
truncLogNormDens(mu, x, sd, threshold)
}
\description{
Log-likelihood of the (two sided) truncated normal distribution
}
