% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTweedy.R
\name{truncNormMLE}
\alias{truncNormMLE}
\title{Computes the conditioal MLE for a truncated normal observation}
\usage{
truncNormMLE(x, threshold, sd = 1)
}
\description{
Computes the conditioal MLE for a truncated normal observation
}
