% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricExampleOptim.R
\name{univTruncNormMLE}
\alias{univTruncNormMLE}
\title{Compute the univariate conditional MLE for truncated normal data}
\usage{
univTruncNormMLE(x, threshold, xsd = 1)
}
\arguments{
\item{x}{the observed truncated normal samples}

\item{threshold}{the selection threshold for the selection rule
\code{x < threshold[1]} or \code{x > threshold[2]}. If a scalar is provided,
then threshold is set to \code{c(-abs(threshold), abs(threshold))}.}

\item{xsd}{the standard deviation of the normal observations}
}
\description{
Compute the univariate conditional MLE for truncated normal data
}
