% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricExampleOptim.R
\name{estimateCensoredParamModel}
\alias{estimateCensoredParamModel}
\title{A function for estimating parametric model 2 (censoring)}
\usage{
estimateCensoredParamModel(x, threshold)
}
\arguments{
\item{x}{the observed data z-scores}

\item{threshold}{the selection threshold for the selection rule abs(x) > threshold}
}
\description{
Parametric model 2 assumes that mu ~ N(0, b) and x ~ N(mu, 1), where x is discarded
if abs(x) < threshold.
}
