% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricExampleOptim.R
\name{estimateTruncParamModel}
\alias{estimateTruncParamModel}
\title{A function for estimating parmetric model 1 (truncation)}
\usage{
estimateTruncParamModel(x, threshold, maxSigma = NULL, nIntPoints = 10^4,
  seed = NULL)
}
\arguments{
\item{x}{the observed z-scores}

\item{threshold}{the selection threshold for the selection rule abs(x) > threshold}

\item{maxSigma}{an optional maximal value for the standard deviation of mu}

\item{nIntPoints}{number of samples to take for numerical integration}

\item{seed}{an optional seed}
}
\description{
A function for estimating the model which assumes that
mu ~ N(0, b) and y ~ TN(mu, 1, threshold).
}
