% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricExampleOptim.R
\name{univTruncNormMLE}
\alias{univTruncNormMLE}
\title{Compute the univariate conditional MLE for truncated normal data}
\usage{
univTruncNormMLE(x, threshold, xsd = 1)
}
\arguments{
\item{x}{the observed truncated normal samples}

\item{threshold}{the threshold for the selection rule abs(x) > threshold}

\item{xsd}{the standard deviation of the normal observations}
}
\description{
Compute the univariate conditional MLE for truncated normal data
}
